requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
filelock==3.7.1
Flask==2.1.3
googlesearch-python==1.0.1
idna==2.10
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
MarkupSafe==2.1.1
mysql-connector==2.2.9
numpy==1.21.4
pandas==1.3.5
python-dateutil==2.8.2
pytz==2022.1
regex==2022.7.9
requests==2.25.1
requests-file==1.5.1
scikit-learn==0.21.3
scipy==1.7.3
six==1.11.0
sklearn==0.0
soupsieve==2.3.2.post1
threadpoolctl==3.1.0
tldextract==3.3.1
typing_extensions==4.3.0
urllib3==1.26.10
Werkzeug==2.2.0
whois==0.9.13
wincertstore==0.2
zipp==3.8.1
